rename slash command
