+++ src/commands/analyze.ts
@@ -1,29 +1,29 @@
 import { SlashCommandBuilder } from 'discord.js'
-import { checkURLType } from '../queues'
+import { checkURL } from '../queues'
 
 export default {
   data: new SlashCommandBuilder()
-    .setName('pdf')
+    .setName('analyze')
     .addStringOption((option) =>
-      option.setName('url').setDescription('URL to PDF file').setRequired(true)
+      option.setName('url').setDescription('URL to report file').setRequired(true)
     )
     .setDescription(
       'Skicka in en årsredovisning och få tillbaka utsläppsdata.'
     ),
 
   async execute(interaction) {
-    console.log('pdf')
+    console.log('analyze')
     const url = interaction.options.getString('url')
     if (!url) {
       await interaction.followUp({
-        content: 'No url provided. Try again with /pdf <url>',
+        content: 'No url provided. Try again with /analyze <url>',
         ephemeral: true,
       })
 
       return
     }
 
-    checkURLType.add('download pdf ' + url.slice(-20), {
+    checkURL.add('download analyze ' + url.slice(-20), {
       url,
     })
