+++ src/workers/parseText.ts
@@ -1,7 +1,7 @@
 import { Worker, Job } from 'bullmq'
 import redis from '../config/redis'
 import OpenAI from 'openai'
-import prompt from '../prompts/parsePDF'
+import prompt from '../prompts/parseReport'
 import { reflectOnAnswer } from '../queues'
 import config from '../config/openai'
 
@@ -17,8 +17,8 @@ class JobData extends Job {
 const worker = new Worker(
   'parseText',
   async (job: JobData) => {
-    const pdfParagraphs = job.data.paragraphs
-    job.log(`Asking AI for following context and prompt: ${pdfParagraphs.join(
+    const { paragraphs } = job.data
+    job.log(`Asking AI for following context and prompt: ${paragraphs.join(
       '\n\n'
     )}
     ${prompt}`)
@@ -26,7 +26,7 @@ const worker = new Worker(
     const stream = await openai.chat.completions.create({
       messages: [
         { role: 'system', content: prompt },
-        { role: 'user', content: pdfParagraphs.join('\n\n') },
+        { role: 'user', content: paragraphs.join('\n\n') },
       ],
       model: 'gpt-4-1106-preview',
       stream: true,
@@ -43,7 +43,7 @@ const worker = new Worker(
     reflectOnAnswer.add('reflect on answer ' + response.slice(0, 20), {
       answer: response,
       url: job.data.url,
-      paragraphs: pdfParagraphs,
+      paragraphs: paragraphs,
     })
 
     return response
